Import dokan-dev/dokan-dotnet e5dfe2a

+++ DokanNet/FileSystemFeatures.cs
@@ -3,7 +3,7 @@
 namespace DokanNet
 {
     [Flags]
-    public enum FileSystemFeatures : uint
+    public enum FileSystemFeatures : long
     {
         None = 0,
         CaseSensitiveSearch = 1,
